Chapter 5 - training models
Mainly linear regression methods. But it does show how you could use linear regression methods to analyze/fit nonlinear models. There is also polynomial regression, which is cool. The main way they tackle under- and over-fitting is through cross-validation and all that. The data is definitely not "cleansed". So, this is where I have an opportunity.